Initial Commit (#6)
* Initial Commit

In app.html added inline styles to clear margin and padding on body. this will allow for easer css manip on all pages.

Added container around form to center the content on the page and add padding

Translate Button: Added styling for disabled state and enabled state with transition between the two. Position is absolute with a negative z index, so the dropdowns cover the button.

Fixed how dropdown chevrons position is calculated, so it will always fit in the container. Before when centering the content it was aligned using "left: Xrem;" which caused variation when centered.

* updates

---------

Co-authored-by: Alexis Raya <aar335@drexel.edu>